Added version control functionality to *.song files
So.. There really isn't a great way to keep version control over
Studio One files.  This is the next best thing IMO.

When the project first gets extracted, the software creates a copy
of the *.song file and rename it *_original.song.  The purpose of this
file is to have something to compare to.  When compared against *.song,
we can determin if the user made any changes to the project since it was
extracted.

What this means is that the next time we download a compressed project
file and extract it, we can test first if there was any changes made to
the local project.  If there was not, then we can just overwrite the
*.song and *_original.song files.  If there was changes, then we can
copy the *.song file and rename it to *_yourversion.song.  The purpose
of this is so that any changes the user made to the local project does
not get overwritten.  Then it gives the user a chance to copy over the
changes they made in *_yourversion.song to the *.song file.

After they copy over their changes to the new file, they can delete the
*_yourversion.song file and compress/upload the new project.

All of this works really easy when only one person makes changes and
uploads at a time. The difficult part becomes when multiple ppl change
the same project at the same time and then try to share.