Changed the compress menu to work with version control
The same things I did to the extract command are also fitted into the
compress command.  It will only compress the main *.song file and it
will copy the *.song file and rename it as the new *_original.song file
so that the software knows it is up to date.