Added Drive support and hash checking
This was a pretty big update.  Drive upload and download ability
was added to the extract menu.  This checks the drive database
hash of the project in question against any local projects to see
if the user already has downloaded the most up-to-date version.
We don't want to spend time downloading if we don't have to!

This also syncs up the audio pools between remote and local projects
as long as there are no detected saved changes on the local project.