Added drive api to compress menu
Now, when compressing a project, it will check the drive for a new
version first.  If you have the same version in the cache as the cloud
does, then it will start compression.

Once finished, it will upload the project to the cloud and update the
hash in the db.json file both remote and locally.