Added open song feature
This feature allows you to open a project, or open the local project as
well as the cloud project, if they are in conflict, at the same time.

When opening a project, this software will create a *_temp.song project
and will open that for the user to modify.  When the user is done and
closes out of studio one, then the software copies *_temp.song to *.song
and deletes the temp song.

We do this to mitigate a specific issue.  We want the user to ONLY open
a project associated with LOF through this program.  This is a
pre-cursor to the dummy clip feature that will be implemented later, as
well as any other future counter measures to prevent conflicts.

The way this works is, a temp song gets created with a different name,
then when the user closes studio one, the temp song gets copied to the
real song and deletes the temp song.  When doing this, studio one loses
track of the filepath to that song.  This prevents the user from opening
the song directly from studio one since it will lose the song path as
soon as it exits.

I hope that makes sense.