Create dummy .wav files
This commit adds dummy *.wav files to the Media dir of each project.

This takes the highest iteration of recording and fills in any open
iterations less-than to create an empty file.  This forces Studio One to
create an audio file that is higher than the highest iteration instead
of trying to fill in lower iterations first.

The reasoning behind this is to prevent as many version control
conflicts as we possibly can.