Added audio file conflict finder
So this basically compares all new audio files against the dummy max
list.  What ever the max iter audio file was, the software compares that
to the newer audio files being downloaded.  If there is two of the same
file, then the software warns the user about this.

It also copies both files to the local project, but renames the old one
to '*_old.wav'.