Check for updates at startup
Before the main menu loop begins, check for updates so that new versions
of the program are installed automatically.

This commit also fixes a bug on MacOS, where calling the `pip` command
would fail. When installed via Homebrew, the `python` and `pip` commands
are installed as `python3` and `pip3`. This commit adds a conditional to
account for the difference in naming.