Remove Slack endpoints from repository
Do not include the Slack endpoint URLs in source code. These URLs are
meant to be kept secret – if the wrong person came across them, they
could theoretically cause some chaos or damage.

Now, the endpoint URLs are obtained out-of-band by prompting the user
directly.

To be safe, all of the webhook endpoint URLs have been regenerated. The
endpoint URLs that previously existed in source code can't really be
erased from the commit history, meaning that someone who was very
motivated could dig up the old endpoint URLs and cause chaos or damage.
We avoid this extremely unlikely situation by regenerating the URLs. The
URLs currently in use have never been committed to the project,
eliminating the possibility of any URLs being found in git or on GitHub.