Added fix for windows and the new 'core'
So for what ever reason, on windows, it couldnt find the 'src' module
from within the 'src' directory.  On mac its fine..

So to fix this, I added a .core.py main program and a .main.py program
all located in the root directory instead of starting from the 'src'
drectory.