Fixed capitalization with module name
So, I had changed 'src/dev.py' to 'src/Dev.py' because of the whole OOP
idea and wanted to keep things uniform.  Unfortunately git did not see
it the same way and does not update capitalization for file names..

Because of this, I will have to write something in the next migration
that changes the name back and forth on next update.  I can't do it in
this update because I do not have the necessary module imports to do so
yet until I can install the 1.2 update first.

Kind of a pickle.

So my solution was to just rename everything back to the lowercase
version for now.  It's not as uniform but at least it works.