Fixed hash bug during new song pulldown
When downloading, opening, saving, then re-uploading a new song in the
same session, the hash never updates for the new compressed file that
was pulled down.  This causes the software to think there is a new
version available on the cloud and does not allow the user to upload
until the next session.

This patch forces the project object to re-hash when it updates.