Patch for Slack Notification Bug
Sometimes if Slack doesn't connect to it's servers right away, it can
return an exception error and crashes the software.  This patch allows
the slack notification class to run up to 5x to connect.  If it fails
all 5x, then the slack message is not sent, however, the software
continues to upload the project.