Added sync function to Database Class
This function is used to sync an orphaned entry with it's remote
counterpart.

When creating a project object with a local project, the entry is
created with ONLY the project name.  The sync function will try and find
an existing entry on the cloud and pair it up with the local project
entry.  If no match is found, the project is left alone.