Fixed error when compressing and uploading
The error occurred after a user compressed and uploaded a project.  If
the user tried to go back to open that project without exiting the
program first, it would crash.  This is because the compress and upload
methods would clean out the temp folder, thus cleaning out the db.json.