Fixed issue with not extracting on upload
The cached audio files were getting wiped out when trying to upload
changes.  This caused the "you created audio files with out your name in
them" warning dialog to pop up every time regardless if you did or not.

This patch also added in options on the project options menu to
 - upload on dirty without opening studio one
 - Download without opening studio one
 - Unlock if exiting a song was error-some (only can unlock yourself)

This patch also addressed the issue of not cleaning out the temp folder.
This clears the temp folder on open and close.