Updated Template
Added new features from Studio One 5 into the template.  This gives the
ability to use multiple console mixes as different scenes.  One scene
for practice and one scene for each member of the band for home use.

This also adds in individual outputs for real-time monitoring through
the software again.  If it doesn't work out or has too much latency, we
will revert back to the old way of doing things.