Rewrite installation section of README
Update the installation instructions with some additional detail, and
with the instructions for each platform broken out into a separate
section.

The setup process for Windows users in particular isn't all that
straight-forward. Even after performing the installation a couple of
times, it's still easy to miss a step.

My hope is that the `README.md` is detailed enough that anyone
attempting an installation will have enough information to know what is
and isn't working.

I also formatted the document, which now complies with the spec. More
information can be found here: [GitHub Flavored Markdown Spec]

[GitHub Flavored Markdown Spec]: https://github.github.com/gfm/