Rewrite Discord.py
This class was created by duplicating the old `Slack.py` class without
modification. When I began to use the new `Discord` class, I started to
realize that the old `Slack` class had some bloat to it. For example,
I noticed there was a some manual HTTP handling and hand-rolled retry
behavior. I also noticed that much of the class was checking
`Dev.isDev()`, and there were a lot of places where the environment was
being checked.

I decided that it would be a good opportunity to refresh the class and
try to strip it down to the bare essentials. For example, I moved a lot
of data into attributes, such as the environment state. This allowed me
to simplify a lot of the environment checking. I also changed the
constructor behavior so that instantiating a new object wouldn't have
the side effect of firing the HTTP request—firing the request was moved
to an instance method.

Speaking of HTTP—I decided to completely punt on that. I found a very
simple `discord-webhooks` library and using that seemed like a much,
much better idea than learning Discord's webhook APIs.

There is still room for improvement, and there's also the chance that
some of what I removed will need to go back in. However, I think this is
a solid step in the right direction, and sets a good precedent for class
design going forward.