Refactor Discord.py
I didn't feel great about the half-refactoring I started in the last
commit, so I decided to keep working with it until I felt it was in a
better place.

Now, the `Discord` class has a much clearer, singular focus. I was able
to push the other concerns out of this class, such as file I/O, Drive
URL string construction, and settings initialization. These things have
been moved to places that I think make sense (with the exception of the
log file I/O, which really needs to be wrapped in its own class, but
that's a different issue entirely).

The `Discord` class now has a predictable and simple API. `Discord`
instances have three main methods which can be used by the caller to
post stuff in the Discord channel. The methods are `post_message`,
`post_link`, and `post_log`. Each behave slightly differently and have
minor parameter differences. This pattern will be easy to continue
developing as the needs of the app change.