Migrate from Slack to Discord
Update the project so that notifications and messages are posted to
their new channel in the Land of Fires Discord server.

Discord has a very similar structure to Slack—both services allow
messages to be created via webhooks. So, the bulk of this migration
comes down to updating webhook URLs and renaming "Slack" to "Discord".