Added migration file and changed settings class
Created the migration 1_4 file to create the keys for the discord
endpoints inside the .settings file.  This migration also removes the
old slack endpoints as well as the 'user' key that was deprecated a bit
ago.

Changed the discord set endpoints function in Settings class to allow
for manual entry.  The automatic way was not finished and needed quite a
bit more work to get running so we are defaulting back to this.